requirements.txt
torch>=2.0.0
torchvision>=0.15.0
gradio>=3.35.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
diffusers>=0.19.0
transformers>=4.30.0
basicsr>=1.4.2
xformers>=0.0.20